Rank-estimation-main/requirements.txt
python-telegram-bot>=20.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
openpyxl>=3.0.0
matplotlib>=3.6.0
flask>=2.3.0

